refactor/PRI-88 Combobox refactor
