+++ vex_app/frontend/src/state/authState.ts
@@ -1,11 +1,11 @@
 // src/state/authState.ts
 import { atom, RecoilState } from 'recoil';
-import persistAtom from './persist'; // persist.tsからimport
+import persistAtom from './persist';
 
 export const authState = atom({
   key: 'isLoginState',
   default: false,
-  effects_UNSTABLE: [persistAtom] // ここでpersistAtomを適用
+  effects_UNSTABLE: [persistAtom] 
 });
 
 // User型の定義が必要です
@@ -15,6 +15,6 @@ interface User {
 
 export const userItemState: RecoilState<User> = atom({
   key: 'userItemState',
-  default: {}, // デフォルト値は適切なUser型のデフォルト値にする
-  effects_UNSTABLE: [persistAtom] // ここでpersistAtomを適用
+  default: {}, 
+  effects_UNSTABLE: [persistAtom] 
 });
